+++ packages/server/__snapshots__/1_es_modules_spec.js
@@ -95,6 +95,7 @@ SyntaxError: /foo/bar/.projects/e2e/lib/fail.js: Unexpected token (2:0)
 > 2 |
     | ^
  @ ./cypress/integration/es_module_import_failing_spec.js 3:0-25
+      [stack trace lines]
 
 This occurred while Cypress was compiling and bundling your test code. This is usually caused by: